requirements.txt
streamlit>=1.34.0
openai>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0
python-dotenv